Added textures.
2D textures and cube maps can be used in shaders by assigning them in a
material.

+++ GameEngine/GameEngine/src/graphics/Shader.cpp
@@ -318,7 +318,7 @@ namespace engine
 
       ShaderTexture texture = {
         _name,
-        static_cast<int>(m_textures.size()),
+        m_textures.size(),
         location
       };
